Tasks/task 5

+++ apps/backend/src/api/tasks/service.ts
@@ -5,22 +5,26 @@ import * as schemas from './schemas.js'
 import { ZodError } from 'zod'
 
 export const getTasks = async (req: Request, res: Response) => {
+  const userId = req.user?.id
+
   const result = await dataSource
     .getRepository(TaskEntity)
-    .createQueryBuilder("repository")
+    .createQueryBuilder("tasks")
+    .where("tasks.userId = :userId", { userId })
     .getMany()
 
   res.status(200).send(result)
 }
 
 export const deleteTaskById = async (req: Request, res: Response) => {
+  const userId = req.user?.id
   const { taskId } = req.params
 
   const deletedRepository = await dataSource
     .createQueryBuilder()
     .delete()
     .from(TaskEntity)
-    .where("id = :taskId", { taskId })
+    .where("id = :taskId AND userId = :userId", { taskId, userId })
     .execute()
 
   if (deletedRepository.affected !== 1) {
@@ -44,13 +48,14 @@ export const editTaskById = async (req: Request, res: Response, next: NextFuncti
     return
   }
 
+  const userId = req.user?.id
   const { taskId } = req.params
 
   const updatedTask = await dataSource
     .createQueryBuilder()
     .update(TaskEntity)
     .set(schemaResult)
-    .where("id = :taskId", { taskId })
+    .where("id = :taskId  AND userId = :userId", { taskId, userId })
     .execute()
 
   if (updatedTask.affected !== 1) {
@@ -74,11 +79,17 @@ export const createTask = async (req: Request, res: Response, next: NextFunction
     return
   }
 
+  const userId = req.user?.id
+  if (!userId) {
+    res.status(401).send("Unauthorized")
+    return
+  }
+
   const createdTask = await dataSource
     .createQueryBuilder()
     .insert()
     .into(TaskEntity)
-    .values([schemaResult])
+    .values([{ ...schemaResult, userId }])
     .returning("*")
     .execute()
